Conflicts:
	Dev/Code/Zerg/Controllers/Community/OrderController.cs

+++ Dev/Ionic/yoopoon/www/page/car/controller/tabCarController.js
@@ -71,23 +71,8 @@ var carlistcount=0;
   };
   $scope.delegateHandler = $ionicSlideBoxDelegate;
 
-    ////全选按钮功能
-//    $scope.start=false;
-//    $scope.allButton=false;
-//    $scope.all=function(){
-//        if($scope.allButton==false){
-//            $scope.start=false;
-//        }if($scope.allButton==true){
-//            $scope.start=true;
-//        }
-//    }
 
-    var el=document.getElementsByTagName("input");
-   var a=el.length;
-    console.log(a);
-    var len=el.length;
-    $scope.allButton=function(){
-        for(var i=0;i<len;i++){
+
 
     //region 全选 单选
     $scope.start=false;//默认未选中
@@ -96,12 +81,15 @@ var carlistcount=0;
     $scope.all=function(c,v){
 
         if(c==true){
+            $scope.choseArr=[""];
             cleanchoseArr();
-            for( i=0;i<$scope.productlist.length;i++)
+            for( j=0;j<$scope.productlist.length;j++)
             {
-                $("#check"+$scope.productlist[i].id).find(":checkbox")[0].checked=true;
-                $scope.choseArr.push($scope.productlist[i].id);
+                cleanchoseArr();
+                $("#check"+$scope.productlist[j].id).find(":checkbox")[0].checked=true;
+                $scope.choseArr.push($scope.productlist[j].id);
             }
+           // document.getElementById('ccsum').removeAttribute('disabled');
            // var valuss=v;
            // $scope.choseArr=valuss;
         }else{
@@ -109,10 +97,13 @@ var carlistcount=0;
             {
                 $("#check"+$scope.productlist[i].id).find(":checkbox").eq(0).prop('checked', false);
             }
+
             $scope.choseArr=[""];
             cleanchoseArr();
+           // document.getElementById('ccsum').setAttribute('disabled');
         }
         allprice();
+
     }
     var cleanchoseArr=function() {
         for(i=0;i<$scope.choseArr.length;i++)
@@ -188,12 +179,15 @@ var carlistcount=0;
 
 
     $scope.decrease=function(id){
-        cartservice.delete(id);
 
         for(j=0;j<$scope.productlist.length;j++){
             if($scope.productlist[j].id==id){
-                $scope.productlist[j].count=  $scope.productlist[j].count-1;
+                if(  $scope.productlist[j].count>1){
+                    $scope.productlist[j].count=  $scope.productlist[j].count-1;
+                    cartservice.delete(id);
+                }
             }
+
         }
         allprice();
 
@@ -203,23 +197,38 @@ var carlistcount=0;
 
     //region 计算总价
     $scope.dprice=0;
+    $scope.sum=0;
+
 
    var allprice=function(){
         var prices=0;
+       $scope.sum=$scope.choseArr.length;
 
        for(i=0;i< $scope.choseArr.length;i++)
        {
            for(j=0;j<$scope.productlist.length;j++){
                if($scope.choseArr[i]==$scope.productlist[j].id){
-                   prices+= parseInt($scope.productlist[j].price * $scope.productlist[j].count);
+                   prices+= parseFloat($scope.productlist[j].price * $scope.productlist[j].count);
+                 //  $scope.sum+=1;
+
                }
+
            }
+       }
+       if( $scope.sum>0){
+           document.getElementById('ccsum').removeAttribute('disabled');
+       }
+       else{
+           // document.getElementById('ccsum').setAttribute('disabled');
+           //  document.getElementById('ccsum').disabled="disabled";
+           $("#ccsum").attr("disabled",true);
+           // $("#ccsum").css("disabled","disabled");
        }
         $scope.dprice=prices;
     }
     //endregion
 
-    //region 编辑
+    //region 删除编辑
 
     $scope.flag={showDelete:false,showReorder:false};
     $scope.items=["Chinese","English","German","Italian","Janapese","Sweden","Koeran","Russian","French"];
